Unify TwoTrackG4e classes and split V0 candidate cfis per channel
Two cleanup steps building on the previous KS/Lambda step-2 commit:

1) Unify ResidualGlobalCorrectionMakerTwoTrackG4e and
   ResidualGlobalCorrectionMakerTwoTrackKPiG4e into a single class.

   The two classes shared >2700 lines of identical code; the K-Pi
   variant was a clone-and-modify with three real additions on top of
   the symmetric J/psi version: per-daughter masses (kaonMass/pionMass),
   a respectTrackOrder flag, and an L1 trigger decision block. All three
   are useful generically (not K-Pi specific), so the right move is to
   keep the more general code and retire the original.

   Concretely, ResidualGlobalCorrectionMakerTwoTrackG4e.cc is replaced
   by the (renamed) K-Pi version with parameter names changed to be
   channel-agnostic:
     kaonMass    -> daughterMass1
     pionMass    -> daughterMass2
     kaonMassErr -> daughterMass1Err
     pionMassErr -> daughterMass2Err
   All new parameters (daughterMass1/2, daughterMass1/2Err, minPairMass,
   maxPairMass, respectTrackOrder, doL1Trigger, l1Results, l1Triggers)
   are made optional with backward-compatible defaults: muon mass for
   the daughters, no pair-mass cut, no track-order constraint, no L1
   filter. So the existing J/psi/Upsilon runner cfgs (which simply
   don't pass any of these) reduce exactly to the previous behaviour.

   ResidualGlobalCorrectionMakerTwoTrackKPiG4e.cc is deleted. The three
   channel-specific cfi clones (KPi for D*, PiPi for KS, ProtonPi for
   Lambda) are updated to instantiate the unified class with the new
   parameter names; their content is otherwise unchanged.

2) Split the generic V0CandidateProducer_cfi into channel-specific
   cfis matching the DstToD0PiCandidateProducer_cfi pattern.

   V0CandidateProducer_cfi.py is replaced by:
     KsToPiPiCandidateProducer_cfi.py        (KS-tuned: both pions, symmetric)
     LambdaToProtonPiCandidateProducer_cfi.py (Lambda-tuned: p+pi, asymmetric)

   The C++ class V0CandidateProducer is unchanged; only the cfi
   wrappers are now per-channel. Runners load the matching cfi.

Validation: all three V0 step-2 channels reproduce identical entry
counts and mass-peak positions before and after the refactor:
  KS     1985 entries, peak at 0.4976 GeV (PDG 0.4976)
  Lambda  807 entries, peak at 1.1162 GeV (PDG 1.1157)
  D*      152 entries, peak at 1.860  GeV (PDG 1.8648)
The legacy J/psi runner cfg loads cleanly under the unified class
(falling back to muon-mass defaults via existsAs<>); a future bit-level
J/psi tree comparison would complete the validation.
